Fix reverse array
Fix unit test

Fix recipient sort in Goerli network

Move addToAddressBook logic from send screen to confirmation screen
